BB8_APP layer
This application enables BB8 to make different motions including moving forward/backward, right/left, rotation of body and rotation of head. These motions are controlled using DC and servo motors. PWM signals for the motors are supplied from the target's timers peripherals. The robot is controlled using a mobile application that is connected to the robot through a Bluetooth module.